# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinTrack Multi-Company Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{companyId}/transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [expense, deposit]
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /companies/{companyId}/transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Reports (wow endpoints)
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, companyId, type, amount, date, description, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        type:
          type: string
          enum: [expense, deposit]
        amount:
          type: number
        description:
          type: string
        notes:
          type: ["string", "null"]
        date:
          type: string
          format: date
        referenceNumber:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TransactionInput:
      type: object
      required: [type, amount, date, description]
      properties:
        categoryId:
          type: integer
        type:
          type: string
          enum: [expense, deposit]
        amount:
          type: number
          minimum: 0.01
        description:
          type: string
          minLength: 1
        notes:
          type: string
        date:
          type: string
          format: date
        referenceNumber:
          type: string
    TransactionUpdate:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [expense, deposit]
        amount:
          type: number
          minimum: 0.01
        description:
          type: string
          minLength: 1
        notes:
          type: string
        date:
          type: string
          format: date
        referenceNumber:
          type: string
    TransactionList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
...